rename header config to basic_types
